+++ src/Integer_Add.cpp
@@ -13,7 +13,7 @@ InfiniteArithmetic::Integer InfiniteArithmetic::Integer::Add(Integer otherOne)
     uint16_t digit = 0;
     uint16_t sum = 0;
 
-    for(size_t i = 0; inThis || inOther; i++)
+    for(int64_t i = 0; inThis || inOther; i++)
     {
         // to calculate sum of digits and carry
         sum = carry;
