+++ src/Integer_Multiply.cpp
@@ -22,8 +22,8 @@ InfiniteArithmetic::Integer InfiniteArithmetic::Integer::Multiply(Integer otherO
     uint16_t sum = 0;
 
     // for result
-    size_t shift = 0;
-    size_t place = 0;
+    int64_t shift = 0;
+    int64_t place = 0;
 
     result.Array.resize(thisOne.Array.size() + otherOne.Array.size() + 1);
